Move xml report to own module
