+++ detekt-report-xml/build.gradle.kts
@@ -1,0 +1,4 @@
+dependencies {
+    implementation(project(":detekt-api"))
+    testImplementation(project(":detekt-test"))
+}
