+++ settings.gradle.kts
@@ -9,6 +9,7 @@ include(
     "detekt-metrics",
     "detekt-parser",
     "detekt-psi-utils",
+    "detekt-report-xml",
     "detekt-rules",
     "detekt-sample-extensions",
     "detekt-test",
